requirements.txt
langchain-core
pydantic
beautifulsoup4
langchain
langgraph
langchainhub
langchain-community
langchain-tavily
langchain-chroma
python-dotenv
pytest
langchain-openai